Refactor/#78: reaction Validation 추가 및 주석 추
